experiments/devops/requirements.txt
pandas>=2.2.0
matplotlib>=3.8.2
numpy>=1.26.4
docker>=7.0.0
gitpython>=3.1.41
pyyaml>=6.0.1
tqdm>=4.66.2
pytest>=7.4.0
pytest-cov>=4.1.0
bandit>=1.7.7
safety>=2.3.5
fastapi>=0.109.2
uvicorn>=0.27.1 